:alien: api: destroy route handlers and switch normal fetch
