+++ src/libs/fetchJson.ts
@@ -1,6 +1,6 @@
 // execute fetch api
-const fetchJson = async <T>(url: string): Promise<T> => {
-  const res = await fetch(url);
+const fetchJson = async <T>(url: string, next?: NextFetchRequestConfig): Promise<T> => {
+  const res = await fetch(url, { next });
   if (!res.ok) {
     throw new Error(`HTTP error! status: ${res.status}`);
   }
